+++ pybop/_dataset.py
@@ -38,6 +38,9 @@ class Dataset:
     def __getitem__(self, key):
         return self.data[key]
 
+    def __setitem__(self, key, value):
+        self.data[key] = value
+
     def __repr__(self):
         """
         Return a string representation of the Dataset instance.